+++ types/types.d.ts
@@ -199,6 +199,7 @@ export interface LoadedTsconfig {
   tags: Map<string, string[]>;
   deprecations: Map<string, string>;
   security: Map<string, SecurityRequirement[]>;
+  requestExamples: Map<string, { value: unknown }>;
   title: string | null;
   version: string | null;
 }